// frontend/pdm-web/src/client/sdk/model/model.type.ts
/**
 * Created by HONG on 8/4/16.
 */
export interface RestfulOptions {
	uriPath: any;
	params?: any;
	querystring?: any;
	header?: any;
	isRemoveServicePath?: boolean;
}

export interface RestfulParamsOptions {
	uriPath: any;
	params: any;
	querystring?: any;
	header?: any;
	isRemoveServicePath?: boolean;
}